separate StakingPoolInfoAndActions
